requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
pydantic==2.3.0
httpx==0.24.1
python-dotenv==1.0.0

# Face matching dependencies
opencv-python-headless
insightface
onnxruntime
numpy
requests
Pillow
